Revert "refactor(suite-common): separate selectors logic"
PARTIALLY reverts commit a28c5a0a
