+++ suite-common/wallet-core/src/index.ts
@@ -1,7 +1,6 @@
 export * from './accounts/accountsActions';
 export * from './accounts/accountsConstants';
 export * from './accounts/accountsMiddleware';
-export * from './accounts/accountsSelectors';
 export * from './accounts/accountsReducer';
 export * from './accounts/accountsThunks';
 export * from './blockchain/blockchainActions';
@@ -15,12 +14,10 @@ export * from './explorer/explorerSelectors';
 export * from './device/deviceActions';
 export * from './device/deviceConstants';
 export * from './device/deviceReducer';
-export * from './device/deviceSelectors';
 export * from './device/deviceThunks';
 export * from './discovery/discoveryActions';
 export * from './discovery/discoveryReducer';
 export * from './discovery/discoveryThunks';
-export * from './discovery/discoverySelectors';
 export * from './discovery/discoveryRunningStateLocks';
 export * from './fees/feesActions';
 export * from './fees/feesReducer';
@@ -35,7 +32,6 @@ export * from './send/composeCancelTransaction/composeCancelTransactionThunk';
 export * from './send/sendFormActions';
 export * from './send/sendFormReducer';
 export * from './send/sendFormThunks';
-export * from './send/sendFormSelectors';
 export * from './send/sendFormTypes';
 export * from './stake/stakeActions';
 export * from './stake/stakeConstants';
@@ -47,7 +43,6 @@ export * from './stake/stakeTypes';
 export * from './transactions/transactionsActions';
 export * from './transactions/transactionsReducer';
 export * from './transactions/transactionsThunks';
-export * from './transactions/transactionsSelectors';
 export * from './settings/walletSettingsActions';
 export * from './settings/walletSettingsConstants';
 export * from './settings/walletSettingsReducer';
